+++ twitch/client.py
@@ -1,13 +1,21 @@
+import os
+from configparser import ConfigParser
+
 from .api import (
     ChannelFeed, Channels, Chat, Collections, Communities, Games, Ingests, Search, Streams, Teams,
     Users, Videos
 )
+
+CONFIG_FILE_PATH = '~/.twitch.cfg'
 
 
 class TwitchClient(object):
     def __init__(self, client_id=None, oauth_token=None):
         self._client_id = client_id
         self._oauth_token = oauth_token
+
+        if not client_id:
+            self._read_credentials_from_file()
 
         self._channel_feed = None
         self._channels = None
@@ -21,6 +29,14 @@ class TwitchClient(object):
         self._teams = None
         self._users = None
         self._videos = None
+
+    def _read_credentials_from_file(self):
+        config = ConfigParser()
+        config.read(os.path.expanduser(CONFIG_FILE_PATH))
+
+        if 'Credentials' in config.sections():
+            self._client_id = config['Credentials'].get('client_id')
+            self._oauth_token = config['Credentials'].get('oauth_token')
 
     @property
     def channel_feed(self):